added multiplicative persistence
